# internal/api/docs/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: SharkAuth API
  version: 0.9.0
  summary: Open-source OAuth 2.1 authorization server with native agent delegation.
  description: "SharkAuth is a single-binary, open-source authentication platform purpose-built\nfor the agent era. It ships\
    \ every primitive a modern auth product needs —\nSSO (SAML + OIDC), passkeys (WebAuthn), TOTP MFA, magic links, social\
    \ login,\norganizations + RBAC, audit log, webhooks, reverse-proxy mode — plus\nnative RFC 8693 token-exchange with full\
    \ agent delegation chains.\n\n### Surfaces\n\nThe API is partitioned into five surfaces. Each lives in its own file under\n\
    `sections/` for readability and tooling. The full spec is the union of the\nfive sections plus this master file's components.\n\
    \n| Section | Path prefix | What |\n|---|---|---|\n| OAuth 2.1 Authorization Server | `/oauth/*`, `/.well-known/*` | Token\
    \ issuance, introspection, revocation, DCR, device flow, token-exchange. |\n| Authentication & Identity | `/api/v1/auth/*`\
    \ | User-facing signup/login/MFA/passkey/magic-link/SSO. |\n| Platform | `/api/v1/organizations`, `/roles`, `/permissions`,\
    \ `/users`, `/sso`, `/api-keys`, `/audit-logs`, `/webhooks`, `/agents` | Multi-tenant org model, RBAC, SSO connections,\
    \ agents, webhooks, audit. |\n| Admin Operations | `/api/v1/admin/*`, `/api/v1/vault/*` | Operator tools: health, config,\
...
  contact:
    name: SharkAuth
    url: https://sharkauth.com
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Session JWT (set by login, refresh on `/oauth/token` refresh grant) or

        admin API key. Send as `Authorization: Bearer <token>`.

        '
    cookieAuth:
      type: apiKey
      in: cookie
      name: shark_session
      description: Set by browser flows (signup/login/social login). HttpOnly + Secure + SameSite=Lax.
    dpopHeader:
      type: apiKey
      in: header
      name: DPoP
      description: 'RFC 9449 proof-of-possession JWT signed by the client''s private key.

        Required on protected resources when the access token has a `cnf.jkt`

        thumbprint claim.

        '
    clientSecretBasic:
      type: http
      scheme: basic
      description: OAuth client credentials via HTTP Basic. Used at `/oauth/token` for confidential clients.
    clientSecretPost:
      type: apiKey
      in: header
      name: Authorization
      description: OAuth client credentials in form-encoded body (`client_id` + `client_secret` fields).
    privateKeyJwt:
      type: oauth2
      description: RFC 7521 client_assertion via signed JWT — for high-assurance clients.
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
...
  schemas:
...
    OrganizationMember:
      type: object
      required:
      - user_id
      - organization_id
      - joined_at
      properties:
        user_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        joined_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/organizations/{id}/members:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - organizations
      operationId: listOrganizationMembers
      summary: List members of an organization
      x-source: internal/api/router.go:376
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrganizationMember'
...
security:
- bearerAuth: []
- cookieAuth: []